// src/app/models/createClient.interface.ts
export interface CreateClientModel{
    nombre_cliente: string;
    documento_cliente: string;
    tipo_documento: string
    correo_cliente: string;
    fecha_nacimiento_cli: string;
    acepto_registro: string;
    terminos_condiciones: string;
    telefono_cliente: string;
    telefono2_cliente: string;
    passwords: string;
  }
  